Add navigation bar to the top and create siteNavitgator
Main navigation bar has been added to the top. This will serve as th primary mechanism
for how the user will navigate the pages of the website. Meanwhile, siteNavigator will serve
as the primary function that will take in the data from the other imported .js files and use
them to operate the changing view tabs of the website