Build out home1 and provide structure for home1, 2, 3
Inserted the first image/icon that the website will be using into home1. An 'imgs' folder was
created in the src folder to contain all images the site uses going forward. On top of this,
in the home.js file, each div will have it's own function that will be called on to add all the
necessary elements and styling