Restructure homePage and overall content container
Added in the website background to the 'content' div instead of the homeContainer. This will
ensure that the background stays the same for any of the tabs on the site. The navigation bar was
also added to the content container, so that everything on the page is now falling onto the
background image